requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
appnope==0.1.4
asttokens==3.0.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
frozenlist==1.8.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jupyter-client==8.8.0
jupyter-core==5.9.1
matplotlib-inline==0.2.1
multidict==6.7.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.4
prompt-toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyjwt==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
requests==2.32.5
schedule==1.2.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
twilio==9.10.3
typing-extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
yarl==1.23.0
